Abstract core logic between cli and gui out
